Add preloading for addons view
