+++ add-ons/package.json
@@ -23,6 +23,7 @@
     "react": "^16.8.6",
     "react-apollo-hooks": "^0.4.5",
     "react-dom": "^16.8.6",
+    "react-router-dom": "^4.2.2",
     "styled-components": "^3.4.10",
     "subscriptions-transport-ws": "^0.9.16"
   },
@@ -31,6 +32,7 @@
     "@types/node": "11.13.0",
     "@types/react": "16.8.13",
     "@types/react-dom": "16.8.3",
+    "@types/react-router-dom": "^4.3.4",
     "@types/styled-components": "^4.1.14",
     "husky": "^1.3.1",
     "lint-staged": "^8.1.4",
