+++ add-ons/src/index.tsx
@@ -1,45 +1,22 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { ApolloProvider } from 'react-apollo-hooks';
 
-import App from './core/App';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
+
+import AddonsViewContent from './core/AddonsViewContent';
+
 import 'fiori-fundamentals/dist/fiori-fundamentals.min.css';
 
-import nestServices from './services/nest';
-import {
-  NotificationsProvider,
-  QueriesProvider,
-  MutationsProvider,
-  SubscriptionsProvider,
-  FiltersProvider,
-  ConfigurationsProvider,
-  LabelsProvider,
-  UrlsProvider,
-} from './services';
+function Preload() {
+  return <div></div>;
+}
 
-import appInitializer from './core/app-initializer';
-import { createApolloClient } from './core/apollo-client';
-
-const Services = nestServices(
-  NotificationsProvider,
-  QueriesProvider,
-  MutationsProvider,
-  FiltersProvider,
-  ConfigurationsProvider,
-  LabelsProvider,
-  UrlsProvider,
-  SubscriptionsProvider,
+ReactDOM.render(
+  <BrowserRouter>
+    <Switch>
+      <Route exact path="/preload" component={Preload} />
+      <Route component={AddonsViewContent} />
+    </Switch>
+  </BrowserRouter>,
+  document.getElementById('root'),
 );
-
-(async () => {
-  await appInitializer.init();
-  const client = createApolloClient();
-  ReactDOM.render(
-    <ApolloProvider client={client}>
-      <Services>
-        <App />
-      </Services>
-    </ApolloProvider>,
-    document.getElementById('root'),
-  );
-})();
